Summary Grid binding ขาดตรง เลือก สินค้า แล้ว อัฟเดท รหัสสินค้า

+++ eoqLab/Models/PurchaseOrderModel.cs
@@ -7,13 +7,15 @@ namespace eoqLab.Models
 {
     public class PurchaseOrderModel
     {
+
         public int MaterialId { get; set; }
         public string MaterialName { get; set; }
         public string project { get; set; }
         public int taskId { get; set; }
         public string description { get; set; }
         public int estimate { get; set; }
-        public double rate { get; set; }
+        public decimal rate { get; set; }
         public string due { get; set; }
+
     }
 }
